+++ src/database/populate/medications.py
@@ -11,39 +11,41 @@ import numpy as np
 
 
 ## MEDICATIONS
-table_medications= '''
-    IF OBJECT_ID('medications') IS NOT NULL
-        DROP TABLE medications;  
-    CREATE TABLE medications (
-        START date,
-        STOP date,
-        PATIENT UNIQUEIDENTIFIER not null,
-        PAYER UNIQUEIDENTIFIER,
-        ENCOUNTER UNIQUEIDENTIFIER,
-        CODE integer,
-        DESCRIPTION nvarchar(150),
-        BASE_COST integer,
-        PAYER_COVERAGE integer,
-        DISPENSES integer,
-        TOTALCOST integer,
-        REASONCODE integer,
-        REASONDESCRIPTION nvarchar(150),
-        primary key (PATIENT, ENCOUNTER, START, CODE)
-    );'''
+def populate_table(reset_table: bool):
+    if reset_table:
+        table_medications= '''
+            IF OBJECT_ID('medications') IS NOT NULL
+                DROP TABLE medications;  
+            CREATE TABLE medications (
+                START date,
+                STOP date,
+                PATIENT UNIQUEIDENTIFIER not null,
+                PAYER UNIQUEIDENTIFIER,
+                ENCOUNTER UNIQUEIDENTIFIER,
+                CODE integer,
+                DESCRIPTION nvarchar(150),
+                BASE_COST integer,
+                PAYER_COVERAGE integer,
+                DISPENSES integer,
+                TOTALCOST integer,
+                REASONCODE integer,
+                REASONDESCRIPTION nvarchar(150),
+                primary key (PATIENT, ENCOUNTER, START, CODE)
+            );'''
 
-database_connection.execute_query(table_medications)
+        database_connection.execute_query(table_medications)
 
-data = pd.read_csv('./database/data/synthea_output/csv/medications.csv')   
-df = pd.DataFrame(data)
-df = df.replace(np.nan,'',regex = True)
-df['START']= pd.to_datetime(df['START'], errors='ignore')
-df['STOP']= pd.to_datetime(df['STOP'], errors='ignore')
+    data = pd.read_csv('./database/data/synthea_output/csv/medications.csv')   
+    df = pd.DataFrame(data)
+    df = df.reset_index(drop=True)
+    df = df.replace(np.nan,'',regex = True)
+    df['START']= pd.to_datetime(df['START'], errors='ignore')
+    df['STOP']= pd.to_datetime(df['STOP'], errors='ignore')
 
-for row in df.itertuples():
     populate_medications = '''
         INSERT INTO medications 
         (START,STOP,PATIENT,PAYER,ENCOUNTER,CODE,DESCRIPTION,BASE_COST,PAYER_COVERAGE,DISPENSES,TOTALCOST,REASONCODE,REASONDESCRIPTION)
         VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?)
         '''
-    database_connection.execute_query(populate_medications, tuple(row[1:]))
+    database_connection.fast_insertion_dataframe(populate_medications, df)
